Comprehensive tests for JsonEventFormatter
This required a bit of rework for the implementation, but the result
is cleaner, in my view.

There's one skipped test, for when an attribute (other than data) is
populated via an integer/Boolean token. We should discuss what we
want to happen in that case.

The extension tests had to change as JsonEventFormatter now
validates that the event it's parsed is valid (for structured mode),
and the ones we were using didn't have a source attribute.

(We will probably want to pepper lots of places with calls to
CloudEvent.Validate, but most of those can come later.)

Signed-off-by: Jon Skeet <jonskeet@google.com>