Remove CloudEventJsonInputFormatter from the ASP.NET Core package
This commit moves it to the sample directory for now. There are
elements we'll want to look at more closely before including it
directly in the package, but anyone can easily look at the code and
include it themselves for now, if it suits their needs.

Means that #75 doesn't block the v2 release.

Signed-off-by: Jon Skeet <jonskeet@google.com>