chore: Target .NET 8.0 (for now) in production packages
This commit *just* adds .NET 8 to the target framework list, in all packages. The code will not build at this point, due to nullability checking. That will be fixed in a later commit, but separating the two makes it easier to review.

Later we may well decided to *build* a .NET 8 target, but not pack it - unless we see concrete benefits from doing so over letting .NET 8 use the existing netstandard2.1 target.

Signed-off-by: Jon Skeet <jonskeet@google.com>